requirements.txt
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
curl_cffi==0.12.0
cycler==0.12.1
fonttools==4.59.0
frozendict==2.4.6
idna==3.10
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
multitasking==0.0.11
nltk==3.9.1
numpy==2.3.1
packaging==25.0
pandas==2.3.1
patsy==1.0.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
protobuf==6.31.1
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2024.11.6
requests==2.32.4
scipy==1.16.0
six==1.17.0
soupsieve==2.7
statsmodels==0.14.5
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vaderSentiment==3.3.2
websockets==15.0.1
yfinance==0.2.65
